move to ts file
